requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
anyio==4.8.0
attrs==25.1.0
audioop-lts==0.2.1
av==14.2.0
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
decorator==5.2.0
discord.py==2.5.2
distlib==0.3.9
dnspython==2.7.0
et_xmlfile==2.0.0
ffmpeg-python==0.2.0
filelock==3.17.0
Flask==3.1.0
fonttools==4.56.0
frozenlist==1.5.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
hatch==1.14.0
hatchling==1.27.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
more-itertools==10.6.0
motor==3.7.0
MouseInfo==0.1.3
multidict==6.1.0
narwhals==1.27.1
numpy==2.2.3
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==6.0.0
pluggy==1.5.0
proglog==0.1.10
propcache==0.3.0
protobuf==5.29.3
ptyprocess==0.7.0
pyarrow==19.0.1
PyAutoGUI==0.9.54
pycparser==2.22
pydeck==0.9.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.1
pymongo==4.11.2
PyMsgBox==1.0.9
pyparsing==3.2.1
pyperclip==1.9.0
pypng==0.20220715.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytweening==1.2.0
pytz==2025.1
pywin32-ctypes==0.2.3
pyzbar==0.1.9
qrcode==8.0
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
streamlit==1.44.1
tenacity==9.0.0
toml==0.10.2
tomli_w==1.2.0
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
trove-classifiers==2025.2.18.16
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
userpath==1.9.2
uv==0.6.1
validators==0.34.0
virtualenv==20.29.2
watchdog==6.0.0
Werkzeug==3.1.3
wikipedia==1.4.0
yarl==1.18.3
yt-dlp==2025.2.19
zstandard==0.23.0
